:zap: Migrate esbuild to bun.build
