+++ package.json
@@ -11,9 +11,10 @@
 	},
 	"type": "module",
 	"scripts": {
-		"dev": "vite dev",
-		"build": "bun esbuild.config.js && vite build",
-		"build:watch": "bun esbuild.config.js && __DEV__=true vite build -w",
+		"dev": "bunx --bun vite",
+		"build:static-script": "bun run build-static-script.js",
+		"build": "bun run build:static-script && bunx --bun vite build",
+		"build:watch": "bun run build:static-script && __WATCH__=true bunx --bun vite build -w",
 		"preview": "vite preview",
 		"test": "npm run test:integration && npm run test:unit",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
@@ -37,7 +38,6 @@
 		"@typescript-eslint/parser": "^6.0.0",
 		"autoprefixer": "^10.4.17",
 		"cssnano": "^6.0.3",
-		"esbuild": "^0.19.11",
 		"eslint": "^8.56.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.35.1",
